Function c509_certificate::generate_signed_c509_cert

pub fn generate_signed_c509_cert(
    tbs_cert: Vec<u8>,
    private_key: &PrivateKey,
) -> Result<C509, Error>

C509 certificate contains 2 parts

TBSCertificate

issuerSignatureValue In order to generate an unsigned C509 certificate, the TBS Certificate must be provided. Then the unsigned C509 certificate will then be used to calculate the issuerSignatureValue.

TBS Certificate

tbs_cert - The TbsCertificate is the TBS Certificate containing

c509CertificateType: A certificate type, whether 0 a natively signed C509 certificate following X.509 v3 or 1 a CBOR re-encoded X.509 v3 DER certificate.

certificateSerialNumber: A unique serial number for the certificate.

issuer: The entity that issued the certificate.

validityNotBefore: The duration for which the Certificate Authority (CA) guarantees it will retain information regarding the certificate’s status on which the period begins.

validityNotAfter: The duration for which the Certificate Authority (CA) guarantees it will retain information regarding the certificate’s status on which the period ends.

subject: The entity associated with the public key stored in the subject public key field.

subjectPublicKeyAlgorithm: The algorithm that the public key is used,

subjectPublicKey: The public key of the subject.

extensions: A list of extensions defined for X.509 v3 certificate, providing additional attributes for users or public keys, and for managing relationships between Certificate Authorities (CAs).

issuerSignatureAlgorithm: The algorithm used to sign the certificate (must be the algorithm uses to create IssuerSignatureValue). Generate a signed C509 certificate.

Arguments

tbs_cert - A cbor encoded TBS certificate. private_key - The private key used to sign the certificate.

Returns

Returns a signed C509 certificate.

Errors

Returns an error if tne sign data cannot be converted to CBOR bytes.