Crate c509_certificate

CBOR Encoded X.509 Certificate (C509 Certificate) library

This crate provides a library for generating C509 Certificates. The function is exposed to Javascript through wasm-bindgen.

Please refer to the C509 Certificate for more information.

Modules

c509

C509 Certificate

c509_algo_identifier

C509 Algorithm Identifier This module handle the AlgorithmIdentifier type where OID does not fall into the table.

c509_attributes

C509 Attributes containing Attribute

c509_big_uint

C509 Unwrapped CBOR Unsigned Bignum (~biguint) Please refer to CDDL Wrapping for unwrapped types.

c509_extensions

C509 Extension as a part of TBSCertificate used in C509 Certificate. Extension fallback of C509 OID extension Given OID if not found in the registered OID table, it will be encoded as a PEN OID. If the OID is not a PEN OID, it will be encoded as an unwrapped OID.

c509_general_names

C509 General Names

c509_issuer_sig_algo

C509 Issuer Signature Algorithm as a part of TBSCertificate used in C509 Certificate.

c509_name

C509 type Name Currently only support natively signed c509 certificate, so all text strings are UTF-8 encoded and all attributeType SHALL be non-negative.

c509_oid

C509 OID provides an encoding and decoding of C509 Object Identifier (OID). Please refer to RFC9090 for OID encoding Please refer to CDDL Wrapping for unwrapped types.

c509_subject_pub_key_algo

C509 Issuer Signature Algorithm as a part of TBSCertificate used in C509 Certificate.

c509_time

C509 Time

signing

ED25519 public and private key implementation.

tables 🔒

A bimap table for bidirectional lookup.

tbs_cert

To Be Sign Certificate (TBS Certificate) module use to construct a C509 certificate.

Functions

generate_signed_c509_cert

C509 certificate contains 2 parts

verify_c509_signature

Verify the signature of a C509 certificate.